instructor-and-writer-tutorial/pyproject.toml
[tool.poetry]
name = "writer-framework-default"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
writer-sdk = "1.6.1"
instructor = "1.7.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
